+++ src/main/java/view/OutputView.java
@@ -8,13 +8,13 @@ import observer.Observer;
 
 public interface OutputView {
 
-    void showQuantityView(int buyQuantity);
+    void showQuantityView(long buyQuantity);
 
 
     void showMyPurchasedLotto(List<Observer> myWholeLotto);
 
 
-    void displayFinalResult(LottoWinningResult result, int buyQuantity);
+    void displayFinalResult(LottoWinningResult result, long buyQuantity);
 
 
 }
